major updates for all files and flask files
