+++ README.md
@@ -1,14 +1,15 @@
-gr_project1
-Game Review Project
+<h2>gr_project1</h2><br>
+<h2>Game Review Project</h2>
 
 
-##Introduction.##
+<h1>Introduction.</h1>
 
-For this project i will be expected to make a Flask Application that will contain at least 2 databases, where i will have to apply C(reate)R(ead)U(pdate)D(elete) methods.
+<p>For this project i will be expected to make a Flask Application that will contain at least 2 databases, where i will have to apply C(reate)R(ead)U(pdate)D(elete) methods.<br>
 The project i have chosen to use as my base topic will be around a games review style, being able to use the front end URL to input data to the tables.
 The user should be able to navigate easily through the application, having a search function and a field to read add update or delete data from the database.
-I will also have all the documentation that has been part of the design process, coding and AGILE methodoligies.
-This will include:
+I will also have all the documentation that has been part of the design process, coding and AGILE methodoligies.<br>
+This will include:</p>
+        <ul>
             -A trello board to keep track of progress and goals
             -A risk assessment to document all the potential threats and issues my project may encounter
             -An ERD and chen diagram to display the functionality and relationships between my database
@@ -16,21 +17,22 @@ This will include:
             -Testing documentation, to prove the functionality of my code
             -A functioning front end flask website
             -A collection of all the files and folders in a GIT repository with no additional unnecessary files
-
-I will be also keeping any changes or updates in a log, to see how the project varied due to issues or compatability with the project criteria.
+        </ul>
+<p> will be also keeping any changes or updates in a log, to see how the project varied due to issues or compatability with the project criteria. </p>
 
 __________________________________________________________________________________________________________________________
 
 
+<h1>Design Ideas</h1><br><br>
 
-In this project, i will set up a database that will hold three tables, one for Games, one for user ID and the other for Reviews.
+<p>In this project, i will set up a database that will hold three tables, one for Games, one for user ID and the other for Reviews.<br><br>
 
-In the games table, it will hold a unique id for the game added, then brief information about the game: ID, name, genre, release date and age rating
+In the games table, it will hold a unique id for the game added, then brief information about the game: ID, title, genre.<br><br>
 
-In the user ID, it will have : ID and User Name
+In the user ID, it will have : ID and User Name<br><br>
 
-In the Review table, there will be an id for the input of review, a space to create a review on the game: ID, ID tied to game ID, name, personal rating 1/10, pros' and cons'.
+In the Review table, there will be an id for the input of review, a space to create a review on the game: Review ID, game ID, (game)title, (user)Nickname, personal rating 1/10, review.<br><br>
 
-The navigation of the tables will be straight-forward and easily readable on the game which you select.
+The navigation of the tables will be straight-forward and easily readable on the game which you select.<br><br>
 
-It will offer you to pick a game, add a game or remove a game. In the reviews table, you will be able to add your own review and see others reviews. You will also be able to sort the games in rating order to see what was best to worst/vice versa.
+It will offer you to pick a game, add a game or remove a game. In the reviews table, you will be able to add your own review and see others reviews. You will also be able to sort the games in rating order to see what was best to worst/vice versa.</p>
