+++ flask_project/create.py
@@ -1,0 +1,5 @@
+from application import db
+
+
+db.drop_all()
+db.create_all()
