+++ flask_project/requirements.txt
@@ -1,0 +1,8 @@
+Flask==2.0.1
+SQLAlchemy==1.4.23
+Flask-SQLAlchemy==2.5.1
+WTForms==2.3.3
+Jinja2==3.0.1
+Flask-WTF==0.15.1
+itsdangerous==2.0.1
+MarkupSafe==2.0.1
